Got rid of warning

+++ mp2/graph.cpp
@@ -155,7 +155,7 @@ vector<int> Graph::dijkstra()
     }
 
     for(size_t i = 0; i < NUM_NODES; i++){
-        if(predecessor[i] == INVALID && i != myNodeID) {
+        if(predecessor[i] == INVALID && i != (size_t)myNodeID) {
             for(size_t j = 0; j < NUM_NODES; j++)
                 valid[i][j] = false;
         }
